backend/requirements.txt
fastapi
uvicorn
gunicorn
ncclient
lxml
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
pyjwt
